[MS-118] Refactor: SMS 서비스 리팩터링
